--- src/main/java/com/vidcentral/api/dto/request/TokenRequest.java
@@ -1,10 +1,0 @@
-package com.vidcentral.api.dto.request;
-
-import lombok.Builder;
-
-@Builder
-public record TokenRequest(
-	String accessToken,
-	String refreshToken
-) {
-}
