+++ src/main/java/com/vidcentral/api/dto/request/auth/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.vidcentral.api.dto.request;
+package com.vidcentral.api.dto.request.auth;
 
 import lombok.Builder;
