+++ src/main/java/com/vidcentral/api/dto/response/auth/LoginResponse.java
@@ -1,4 +1,4 @@
-package com.vidcentral.api.dto.response;
+package com.vidcentral.api.dto.response.auth;
 
 import lombok.Builder;
