+++ src/main/java/com/vidcentral/api/dto/response/member/MemberInfoResponse.java
@@ -1,4 +1,4 @@
-package com.vidcentral.api.dto.response;
+package com.vidcentral.api.dto.response.member;
 
 import lombok.Builder;
